+++ negocio/ventas/metodosDePago/IMetodoDePago.java
@@ -2,5 +2,5 @@ package tpo.negocio.ventas.metodosDePago;
 
 public interface IMetodoDePago {
   public String getMetodo();
-  public double getMultiplicador();
+  public float getMultiplicador();
 }
